Moved e_clouds to .h file
